Move exif package to common
