+++ QKSMS/src/main/java/com/moez/QKSMS/common/exif/CountedDataInputStream.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.moez.QKSMS.exif;
+package com.moez.QKSMS.common.exif;
 
 import java.io.EOFException;
 import java.io.FilterInputStream;
