+++ QKSMS/src/main/java/com/moez/QKSMS/common/exif/ExifInterface.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.moez.QKSMS.exif;
+package com.moez.QKSMS.common.exif;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
