+++ QKSMS/src/main/java/com/moez/QKSMS/common/exif/IfdData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.moez.QKSMS.exif;
+package com.moez.QKSMS.common.exif;
 
 import java.util.HashMap;
 import java.util.Map;
